+++ src/Exception/RequestException.php
@@ -18,9 +18,6 @@ class RequestException extends TransferException
     /** @var ResponseInterface */
     private $response;
 
-    /** @var bool */
-    private $throwImmediately = false;
-
     public function __construct(
         $message = '',
         RequestInterface $request,
@@ -122,28 +119,5 @@ class RequestException extends TransferException
             throw new \InvalidArgumentException('You cannot set the emitted '
                 . 'error value to false.');
         }
-    }
-
-    /**
-     * Sets whether or not parallel adapters SHOULD throw the exception
-     * immediately rather than handling errors through asynchronous error
-     * handling.
-     *
-     * @param bool $throwImmediately
-     *
-     */
-    public function setThrowImmediately($throwImmediately)
-    {
-        $this->throwImmediately = $throwImmediately;
-    }
-
-    /**
-     * Gets the setting specified by setThrowImmediately().
-     *
-     * @return bool
-     */
-    public function getThrowImmediately()
-    {
-        return $this->throwImmediately;
     }
 }