+++ tests/PoolTest.php
@@ -7,6 +7,11 @@ use GuzzleHttp\Pool;
 use GuzzleHttp\Ring\Client\MockAdapter;
 use GuzzleHttp\Ring\Future;
 use GuzzleHttp\Subscriber\History;
+use GuzzleHttp\Event\BeforeEvent;
+use GuzzleHttp\Event\CompleteEvent;
+use GuzzleHttp\Event\ErrorEvent;
+use GuzzleHttp\Message\Response;
+use GuzzleHttp\Subscriber\Mock;
 
 class PoolTest extends \PHPUnit_Framework_TestCase
 {
@@ -111,5 +116,52 @@ class PoolTest extends \PHPUnit_Framework_TestCase
         });
 
         return new Client(['adapter' => new MockAdapter($future)]);
+    }
+
+    public function testBatchesRequests()
+    {
+        $client = new Client();
+        $responses = [
+            new Response(301, ['Location' => 'http://foo.com/bar']),
+            new Response(200),
+            new Response(200),
+            new Response(404)
+        ];
+        $client->getEmitter()->attach(new Mock($responses));
+        $requests = [
+            $client->createRequest('GET', 'http://foo.com/baz'),
+            $client->createRequest('HEAD', 'http://httpbin.org/get'),
+            $client->createRequest('PUT', 'http://httpbin.org/put'),
+        ];
+
+        $a = $b = $c = 0;
+        $result = Pool::batch($client, $requests, [
+            'before'   => function (BeforeEvent $e) use (&$a) { $a++; },
+            'complete' => function (CompleteEvent $e) use (&$b) { $b++; },
+            'error'    => function (ErrorEvent $e) use (&$c) { $c++; },
+        ]);
+
+        $this->assertEquals(4, $a);
+        $this->assertEquals(2, $b);
+        $this->assertEquals(1, $c);
+        $this->assertCount(3, $result);
+
+        // The first result is actually the second (redirect) response.
+        $this->assertSame($responses[1], $result[0]);
+        // The second result is a 1:1 request:response map
+        $this->assertSame($responses[2], $result[1]);
+        // The third entry is the 404 RequestException
+        $this->assertSame($responses[3], $result[2]->getResponse());
+    }
+
+    /**
+     * @expectedException \InvalidArgumentException
+     * @expectedExceptionMessage Invalid event format
+     */
+    public function testBatchValidatesTheEventFormat()
+    {
+        $client = new Client();
+        $requests = [$client->createRequest('GET', 'http://foo.com/baz')];
+        Pool::batch($client, $requests, ['complete' => 'foo']);
     }
 }